Implement rudimentary static site generator
Run `php gnerate.php` to generate the html site in the generated
directory

I tried a couple of off the shelf static site generators (sculpin and statie)
but had trouble configuring them, so in the end it seemed easier to just
pull in some libraries and write my own minimal generator in one PHP
file.

I might end up replacing it with a Symfony app with a script to save
every page.